% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SARA_R.R
\name{SARA}
\alias{SARA}
\title{Process SARA report files into dataframes and save to a spreadsheet}
\usage{
SARA()
}
\value{
update from last year's files
}
\description{
Depends on having files in the rawdata subdirectory
Will crash if file format incorrect
}
