% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SARA_readwrite.R
\name{SARA_readwrite}
\alias{SARA_readwrite}
\title{SARA report files into dataframes}
\usage{
SARA_readwrite(dbg = FALSE)
}
\value{
update from last year's files
}
\description{
SARA report files into dataframes
}
