% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SARA_readwrite.R
\name{SARA_readwrite}
\alias{SARA_readwrite}
\title{Actual function that reads in existing Tier 3 datafiles (results) in data subdirectory, then
appends the results from the rawdata subdirectory}
\usage{
SARA_readwrite(dbg = FALSE)
}
\value{
update from last year's files
}
\description{
Draft
}
