% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SARA_readwrite_form.R
\name{SARA_readwrite_form}
\alias{SARA_readwrite_form}
\title{SARA report files into dataframes}
\usage{
SARA_readwrite_form()
}
\value{
update from last year's files
}
\description{
SARA report files into dataframes
}
